Reduce coupling with systemd in extras scripts

+++ extras/linux_generic/install.sh
@@ -184,8 +184,9 @@ if (( uninstall )); then
         rm -fv /etc/profile.d/unum.sh || :
         rm -fv /etc/systemd/system/unum.service || :
         rm -fv /usr/bin/minim-config || :
-        cp -f /etc/default/hostapd.pre-unum /etc/default/hostapd || :
-        cp -f /etc/default/dnsmasq.pre-unum /etc/default/dnsmasq || :
+        cp -fv /etc/default/hostapd.pre-unum /etc/default/hostapd || :
+        cp -fv /etc/default/dnsmasq.pre-unum /etc/default/dnsmasq || :
+        cp -fv /etc/dhcpcd.conf.pre-unum /etc/dhcpcd.conf || :
     else
         echo "did not uninstall unum, exiting without making any changes"
     fi
@@ -279,25 +280,34 @@ fi
 echo "unum config file is $install_etc_dir/config.json"
 cp -f "$dist_dir/etc/opt/unum/config.json" "$install_etc_dir/config.json"
 
-# Add minim-config utility to /usr/bin to avoid any PATH issues.
-declare mod_check
+# Handle all-in-one related installation.
+# Expects hostapd, dnsmasq, and /etc/default to be used
 if (( install_aio )); then
     ln -sf "$extras_dir/sbin/minim-config" "/usr/bin/minim-config"
     echo "installed minim-config: /usr/bin/minim-config"
 
-    mod_check=`awk '/^DNSMASQ_OPTS=.*unum.*$/ { print "1" }' /etc/default/dnsmasq`
-    if [[ -z "$mod_check" ]]; then
-        cp -f /etc/default/dnsmasq /etc/default/dnsmasq.pre-unum
-        echo 'DNSMASQ_OPTS="--conf-file=/etc/opt/unum/dnsmasq.conf"' >> /etc/default/dnsmasq
-        echo "updated /etc/default/dnsmasq to use generated config"
+    declare mod_check
+    if [[ -f "/etc/default/dnsmasq" ]]; then
+        mod_check=$(awk '/^DNSMASQ_OPTS=.*unum.*$/ { print "1" }' /etc/default/dnsmasq)
+        if [[ -z "$mod_check" ]]; then
+            cp -f /etc/default/dnsmasq /etc/default/dnsmasq.pre-unum
+            echo 'DNSMASQ_OPTS="--conf-file='"$install_etc_dir"'/dnsmasq.conf --log-facility='"$install_var_dir"'/dnsmasq.log"' >> /etc/default/dnsmasq
+            echo "updated /etc/default/dnsmasq to use generated config"
+        fi
+    else
+        echo "warning: did not automatically modify /etc/default/dnsmasq (it does not exist!)"
     fi
 
-    mod_check=`awk '/^DAEMON_CONF.*unum.*$/ { print "1" }' /etc/default/hostapd`
-    if [[ -z "$mod_check" ]]; then
-        cp -f /etc/default/hostapd /etc/default/hostapd.pre-unum
-        sed -i -E 's:^[#]?DAEMON_CONF.*:DAEMON_CONF="/etc/opt/unum/hostapd-phy0.conf":' /etc/default/hostapd
-        echo 'DAEMON_OPTS="-dd -t -f /var/opt/unum/hostapd.log"' >> /etc/default/hostapd
-        echo "updated /etc/default/hostapd to use generated config"
+    if [[ -f "/etc/default/hostapd" ]]; then
+        mod_check=$(awk '/^DAEMON_CONF.*unum.*$/ { print "1" }' /etc/default/hostapd)
+        if [[ -z "$mod_check" ]]; then
+            cp -f /etc/default/hostapd /etc/default/hostapd.pre-unum
+            sed -i -E 's:^[#]?DAEMON_CONF.*:DAEMON_CONF="'"$install_etc_dir"'/hostapd-phy0.conf":' /etc/default/hostapd
+            echo 'DAEMON_OPTS="-dd -t -f '"$install_var_dir"'/hostapd.log"' >> /etc/default/hostapd
+            echo "updated /etc/default/hostapd to use generated config"
+        fi
+    else
+        echo "warning: did not automatically modify /etc/default/hostapd (it does not exist!)"
     fi
 fi
